Removed IDriveTrain PracticeDrivetrain, and AbstractDrivetrain
No longer trying to run two drivetrains with this project,
