--- src/main/java/frc/robot/subsystems/IDriveTrain.java
@@ -1,48 +1,0 @@
-/*----------------------------------------------------------------------------*/
-/* Copyright (c) 2018-2019 FIRST. All Rights Reserved.                        */
-/* Open Source Software - may be modified and shared by FRC teams. The code   */
-/* must be accompanied by the FIRST BSD license file in the root directory of */
-/* the project.                                                               */
-/*----------------------------------------------------------------------------*/
-
-package frc.robot.subsystems;
-
-import edu.wpi.first.wpilibj.Encoder;
-import edu.wpi.first.wpilibj.drive.DifferentialDrive;
-import edu.wpi.first.wpilibj2.command.Subsystem;
-/**
- * Add your docs here.
- */
-public interface IDriveTrain extends Subsystem{
-
-
- 
-  /**
-   * Drives the robot using arcade controls.
-   *
-   * @param fwd the commanded forward movement
-   * @param rot the commanded rotation
-   */
-  public void arcadeDrive(double fwd, double rot) ;
-
-  /**
-   * Resets the drive encoders to currently read a position of 0.
-   */
-  public void resetEncoders(); 
-
-
-  /**
-   * Sets the max output of the drive.  Useful for scaling the drive to drive more slowly.
-   *
-   * @param maxOutput the maximum output to which the drive will be constrained
-   */
-  public void setMaxOutput(double maxOutput); 
-
-  public DifferentialDrive getDrive();
-
-  /**
-   * Moves the drivetrain forward that many inches
-   * @param Inches
-   */
-  public void moveTo(double Inches);
-}
